mult/compress 的 MPO 结果一律返回 CanonicalIMPO；实输入复 leading vector 对齐 MPSKit；改为依赖 FiniteMPSAlgorithms
- compress/mult（含 naive 兜底与 D≥ 短路）：MPO 结果一律 CanonicalIMPO 且
  ismixedcanonical，不再透出 DenseIMPO；hadamard 无 MPO 路径不受影响
- 复环境提升（MPSKit 对齐）：环境张量按 eigsolve 的实际 eltype 存放（实输入下
  融合转移的 leading vector 可为复），通道随 _promote_scalar 升为复算术；放宽
  push_env_left/right、_mapAC/_mapC/_galerkin/_ring_overlap 的单标量类型签名；
  gaugefix! 对实链复本征对取实部作加速初值；构造融合转移 ±2i 的专用测试固化
- 依赖 FiniteMPSAlgorithms：删除本包 tensorops 层（5 文件）与重复测试
  （test/tensorops，158 条）；30 个名字显式 import + 再导出，对外 API 不变；
  同秩 N 维 Base.kron 扩展随依赖可用（arithmetics rank-3 zip 还原原写法）
- 修复 MPSKit 新版导出 PeriodicVector 造成的测试作用域歧义（finite_t 加限定）

+++ src/utility.jl
@@ -30,7 +30,7 @@ Base.lastindex(p::PeriodicVector) = length(p.data)
 Base.iterate(p::PeriodicVector, args...) = iterate(p.data, args...)
 Base.similar(p::PeriodicVector{T}) where {T} = PeriodicVector(similar(p.data))
 Base.copy(p::PeriodicVector) = PeriodicVector(copy(p.data))
-Base.circshift(p::PeriodicVector, n) = PeriodicVector(circshift(p.data, n))
+Base.circshift(p::PeriodicVector, n::Integer) = PeriodicVector(circshift(p.data, n))
 Base.:(==)(a::PeriodicVector, b::PeriodicVector) = a.data == b.data
 _parent(p::PeriodicVector) = p.data
